added a mouse bites (breakable point between PCBs) footprint
their dimensions are based on the mouse bites component[1] on a board[2]
I have had manufactured already.

I hate creating circles/archs in kicad, so this component contains two
half-circles on Edge.Cuts at the ends of the mouse bite row. Edge.Cuts
is an "illegal" layer for components, so you can not create it in the
editor and if you try editing the archs it gives a warning that it
defaults to F.SilkS. it was created by using F.SilkS when creating them
and manually changing it to Edge.Cuts with a text-editor, and it seems
to work, but I have no guarantee that it will work in future versions.

[1] https://github.com/thebiguno/microcontroller-projects/blob/47a4a5314f2d2cc3db8d2db4c6288e78091556d9/lib/kicad/modules/custom.mod#L6778-L6807
[2] https://github.com/thebiguno/microcontroller-projects/tree/47a4a5314f2d2cc3db8d2db4c6288e78091556d9/projects/temp